group resource roles

+++ assign_resource_role.py
@@ -2,7 +2,7 @@ import click
 import sys
 import json
 import os
-from utils import show_progress, get_env_info, poll_for_status, get_cdp_env_crn
+from utils import show_progress, get_env_info, poll_for_status, get_cdp_env_crn, get_resource_crn
 from cdpv1sign import generate_headers
 import requests_ops
 import requests
@@ -17,24 +17,22 @@ def dump_assign_group_resource_role_json(cdp_env_crn, cdp_role, group_name, json
     return assign_group_role_json
 
 
-# def dump_assign_user_resource_role_json(cdp_env_crn, cdp_role, user, json_skel):
-#     assign_group_role_json = dict(json_skel)
-#     resource_role_crn = f"{requests_ops.DEFAULT_IAM_CRN}:resourceRole:{cdp_role}"
-#     assign_group_role_json["groupName"] = user
-#     assign_group_role_json["resourceRoleCrn"] = resource_role_crn
-#     assign_group_role_json["resourceCrn"] = cdp_env_crn
-#     return assign_group_role_json
-
-
-# def dump_unassign_group_resource_role_json(cred_info, json_skel):
-#     assign_group_role_json = dict(json_skel)
-#     assign_group_role_json["credentialName"] = cred_info["credential_name"]
-#     return assign_group_role_json
+def dump_unassign_group_resource_role_json(group, cdp_env_crn, cdp_role, json_skel):
+    resource_role_crn = f"{requests_ops.DEFAULT_IAM_CRN}:resourceRole:{cdp_role}"
+    assign_group_role_json = dict(json_skel)
+    assign_group_role_json["resourceRoleCrn"] = resource_role_crn
+    assign_group_role_json["resourceCrn"] = cdp_env_crn
+    assign_group_role_json["groupName"] = group
+    return assign_group_role_json
 
 
 @click.command()
 @click.option("--dryrun/--no-dryrun", default=True)
-@click.option("--action", type=click.Choice(["assign-role-to-group"]), required=True)
+@click.option(
+    "--action",
+    type=click.Choice(["assign-role-to-group", "unassign-role-from-group"]),
+    required=True,
+)
 @click.option(
     "--env",
     type=click.Choice(["lab", "test", "dev", "acc", "prod"]),
@@ -71,30 +69,19 @@ def main(dryrun, env, cdp_env_name, action, json_skel):
                     cdp_env_crn, role, group, assign_group_role_json_skel
                 )
                 action_url = f"{env_url}/assignGroupResourceRole"
-            # elif action == "delete-cred":
-            #     click.echo(f"==============Deleting credential {cred_name}==============")
-            #     cdp_assign_group_role_json = dump_unassign_group_resource_role_json(
-            #         cred_info, assign_group_role_json_skel
-            #     )
-            #     action_url = f"{env_url}/deleteCredential"
+            elif action == "unassign-role-from-group":
+                click.echo(
+                    f"==============Unassigning role {role} from group {group}=============="
+                )
+                cdp_assign_group_role_json = dump_unassign_group_resource_role_json(
+                    group, cdp_env_crn, role, assign_group_role_json_skel
+                )
+                action_url = f"{env_url}/unassignGroupResourceRole"
 
-        click.echo("-------------------Generated JSON-----------------------------")
-        print(json.dumps(cdp_assign_group_role_json, indent=4, sort_keys=True))
-        click.echo("--------------------------------------------------------------")
-
-        if not dryrun:
-            response = requests_ops.send_http_request(
-                srv_url=action_url,
-                req_type="post",
-                data=cdp_assign_group_role_json,
-                headers=generate_headers("POST", action_url),
-            )
-
-            click.echo(f"Waiting for {action} on role {role}")
             if action == "assign-role-to-group":
                 elem_present = True
-            # elif action == "delete-cred":
-            #     elem_present = False
+            elif action == "unassign-role-from-group":
+                elem_present = False
 
             elem_search_info = {
                 "root_index": "resourceAssignments",
@@ -109,6 +96,20 @@ def main(dryrun, env, cdp_env_name, action, json_skel):
                 data={"groupName": group},
                 elem_search_info=elem_search_info,
             )
+
+        click.echo("-------------------Generated JSON-----------------------------")
+        print(json.dumps(cdp_assign_group_role_json, indent=4, sort_keys=True))
+        click.echo("--------------------------------------------------------------")
+
+        if not dryrun:
+            response = requests_ops.send_http_request(
+                srv_url=action_url,
+                req_type="post",
+                data=cdp_assign_group_role_json,
+                headers=generate_headers("POST", action_url),
+            )
+
+            click.echo(f"Waiting for {action} on role {role}")
             click.echo(f"Action {action} on cdp group {group} assigning role {role} DONE")
             # dumping file so that Gitlab will back it up
             with open(f"{group}_{role}.json", "w", encoding="utf-8") as f:
